--- app/src/main/java/br/senac/pi4pokemon/MainActivity.kt
@@ -1,11 +1,0 @@
-package br.senac.pi4pokemon
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
