+++ app/src/main/java/br/senac/pi4pokemon/model/Produto.kt
@@ -1,0 +1,14 @@
+package br.senac.pi4pokemon.model
+
+data class Produto(
+	val preco: String,
+	val updatedAt: String,
+	val imagem: String,
+	val createdAt: String,
+	val nome: String,
+	val id: Int,
+	val categoriaId: Int,
+	val quantidade: String,
+	val deletedAt: String,
+	val descricao: String,
+)
