+++ app/src/main/java/br/senac/pi4pokemon/views/NewAdressActivity.kt
@@ -1,11 +1,12 @@
-package br.senac.pi4pokemon
+package br.senac.pi4pokemon.views
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import br.senac.pi4pokemon.R
 
-class RegisterActivity : AppCompatActivity() {
+class NewAdressActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register)
+        setContentView(R.layout.activity_new_adress)
     }
 }
