+++ app/src/main/java/br/senac/pi4pokemon/views/ProductViewActivity.kt
@@ -1,11 +1,12 @@
-package br.senac.pi4pokemon
+package br.senac.pi4pokemon.views
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import br.senac.pi4pokemon.R
 
-class NewAdressActivity : AppCompatActivity() {
+class ProductViewActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_new_adress)
+        setContentView(R.layout.activity_product_view)
     }
 }
