+++ app/src/main/res/layout/activity_bottom_navigation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/pokeBlueFundo"
-    tools:context=".BottomNavigation">
+    tools:context=".views.BottomNavigation">
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottomNavigationView"
