+++ app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/banner_menu_02"
-    tools:context=".LoginActivity">
+    tools:context=".views.LoginActivity">
 
     <ImageView
         android:id="@+id/imageView"
