+++ app/src/main/res/layout/activity_new_adress.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/pokeBlueFundo"
-    tools:context=".NewAdressActivity">
+    tools:context=".views.NewAdressActivity">
 
     <TextView
         android:id="@+id/textView2"
