+++ app/src/main/res/layout/activity_product_view.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/pokeBlueFundo"
-        tools:context=".ProductViewActivity">
+        tools:context=".views.ProductViewActivity">
 
         <TextView
         android:id="@+id/textView3"
@@ -46,16 +46,16 @@
         app:layout_constraintTop_toBottomOf="@+id/textView3"
         app:srcCompat="@drawable/star_full" />
 
-    <ImageView
-        android:id="@+id/imageView14"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:layout_marginTop="8dp"
-        app:layout_constraintStart_toEndOf="@+id/imageView13"
-        app:layout_constraintTop_toBottomOf="@+id/textView3"
-        app:srcCompat="@drawable/star_full" />
+        <ImageView
+            android:id="@+id/imageView14"
+            android:layout_width="24dp"
+            android:layout_height="24dp"
+            android:layout_marginTop="8dp"
+            app:layout_constraintStart_toEndOf="@+id/imageView13"
+            app:layout_constraintTop_toBottomOf="@+id/textView3"
+            app:srcCompat="@drawable/star_full" />
 
-    <ImageView
+        <ImageView
         android:id="@+id/imageView15"
         android:layout_width="24dp"
         android:layout_height="24dp"
