+++ app/src/main/res/layout/activity_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/banner_menu_02"
-    tools:context=".RegisterActivity">
+    tools:context=".views.RegisterActivity">
 
     <EditText
         android:id="@+id/editTextTextPersonName3"
